requirements.txt
alpaca-py==0.8.2
asttokens==2.2.1
backcall==0.2.0
certifi==2023.5.7
charset-normalizer==3.1.0
comm==0.1.3
debugpy==1.6.7
decorator==5.1.1
executing==1.2.0
idna==3.4
importlib-metadata==6.7.0
ipykernel==6.24.0
ipython==8.12.2
jedi==0.18.2
jupyter-client==8.3.0
jupyter-core==5.3.1
matplotlib-inline==0.1.6
msgpack==1.0.5
nest-asyncio==1.5.6
numpy==1.24.4
packaging==23.1
pandas==1.5.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.8.0
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==1.10.11
Pygments==2.15.1
python-dateutil==2.8.2
pytz==2023.3
pyzmq==25.1.0
requests==2.31.0
six==1.16.0
sseclient-py==1.7.2
stack-data==0.6.2
tornado==6.3.2
traitlets==5.9.0
typing-extensions==4.7.1
urllib3==2.0.3
wcwidth==0.2.6
websockets==10.4
zipp==3.15.0
